Added all 0.7.0 binaries and fixed a couple issues.
Now others should be able to create builds easily since I'm committing
all of the *.dll and *.exe binaries.  I'm old school and I didn't want
to do it but for Gow to live on it needs to be easy to build and
enhance.

I am committing the old 0.7.0 binaries for now.  The next commit or
two will include all of the binaries.

I also updated the version numbers to 0.7.0 to reflect the version.